requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2==2.9.10
psycopg2-binary==2.9.9
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
uritemplate==4.2.0
